system_metrics_collector/requirements.txt
python-dotenv
psutil
prometheus_client
pytz
requests
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc